refactor(spf): relocate pure helpers surfaced by load-segments refactor
Three relocations, all surfaced as pure-helper sniffs by the prior
load-segments reactor refactor:

1. Bandwidth estimation (EWMA + state) → network/.

   bandwidth-estimator.ts + ewma.ts (and their tests) move from
   media/abr/ to network/. EWMA-based bandwidth estimation is a
   network-throughput concern, not a media concern — its placement
   in media/abr/ pre-dated the layer model. quality-selection.ts
   stays in media/abr/ (it consumes a `number`, not BandwidthState
   directly; no media → network dep introduced).

2. fetchStream + createTrackedFetch → network/fetch.ts.

   Two stream-shaped fetch helpers that lived in load-segments.ts
   despite having no core deps. fetchStream is the eager-connect /
   lazy-iterate sibling to fetchResolvableStream (added with a JSDoc
   note explaining the timing difference). createTrackedFetch's API
   simplifies: no longer takes a Signal<BandwidthState>; instead
   accumulates EWMA state internally and notifies via onSample
   callback. Eliminates the local throughput signal in
   loadVideoSegments — state.bandwidthState is now the sole external
   accumulator.

3. segmentStartFor → media/buffer/forward-buffer.ts (renamed to
   segmentStartForTime).

   Pure segment-locator helper, same domain as the existing
   forward-buffer's getSegmentsToLoad / calculateForwardFlushPoint.
   Signature widened to take segments directly (not whole tracks)
   so it composes with `bufferedSegments`-shaped inputs too. Renamed
   for clarity — segmentStartFor read ambiguously.
