+++ packages/spf/src/network/ewma.ts
