+++ packages/spf/src/playback/behaviors/dom/tests/load-segments-track-switch.test.ts
@@ -1,8 +1,8 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import type { ContextSignals, StateSignals } from '../../../../core/composition/create-composition';
 import { signal } from '../../../../core/signals/primitives';
-import type { BandwidthState } from '../../../../media/abr/bandwidth-estimator';
 import type { MaybeResolvedPresentation, Presentation, VideoSelectionSet } from '../../../../media/types';
+import type { BandwidthState } from '../../../../network/bandwidth-estimator';
 import { createSourceBufferActor, type SourceBufferActor } from '../../../actors/dom/source-buffer';
 import type { SegmentLoadingContext, SegmentLoadingState } from '../load-segments';
 import { loadVideoSegments } from '../load-segments';
